requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-filter==2.3.0
requests==2.24.0
django-cors-headers==3.4.0
uwsgi==2.0.19
dj-database-url==0.5.0
pytest-django==3.9.0
coverage==5.2
pytest-cov==2.10
requests-mock